Move stream mapping to instrument config files
