+++ src/beamlime/config/instruments/bifrost.py
@@ -19,6 +19,9 @@ from scippnexus import NXdetector
 
 from beamlime.handlers.detector_data_handler import get_nexus_geometry_filename
 from beamlime.handlers.workflow_manager import processor_factory
+from beamlime.kafka import InputStreamKey, StreamLUT, StreamMapping
+
+from ._ess import make_common_stream_mapping_inputs, make_dev_stream_mapping
 
 
 def _to_flat_detector_view(da: sc.DataArray) -> sc.DataArray:
@@ -144,3 +147,27 @@ source_to_key = {
 f144_attribute_registry = {
     'detector_rotation': {'units': 'deg'},
 }
+
+
+def _make_bifrost_detectors() -> StreamLUT:
+    """
+    Bifrost detector mapping.
+
+    Input keys based on
+    https://confluence.ess.eu/display/ECDC/Kafka+Topics+Overview+for+Instruments
+    """
+    # Source names have the format `arc=[0-4];triplet=[0-8]`.
+    return {
+        InputStreamKey(
+            topic='bifrost_detector', source_name=f'arc={arc};triplet={triplet}'
+        ): f'arc{arc}_triplet{triplet}'
+        for arc in range(5)
+        for triplet in range(9)
+    }
+
+
+stream_mapping_dev = make_dev_stream_mapping('bifrost')
+stream_mapping = StreamMapping(
+    **make_common_stream_mapping_inputs(instrument='bifrost'),
+    detectors=_make_bifrost_detectors(),
+)
