+++ src/beamlime/config/instruments/dream.py
@@ -7,6 +7,10 @@ import sciline
 import scipp as sc
 from ess.reduce.streaming import StreamProcessor
 
+from beamlime.kafka import InputStreamKey, StreamLUT, StreamMapping
+
+from ._ess import make_common_stream_mapping_inputs, make_dev_stream_mapping
+
 
 def _get_mantle_front_layer(da: sc.DataArray) -> sc.DataArray:
     return (
@@ -129,3 +133,32 @@ source_to_key = {
     'monitor1': RawMon1,
     'monitor2': RawMon2,
 }
+
+
+def _make_dream_detectors() -> StreamLUT:
+    """
+    Dream detector mapping.
+
+    Input keys based on
+    https://confluence.ess.eu/display/ECDC/Kafka+Topics+Overview+for+Instruments
+    """
+    mapping = {
+        'bwec': 'endcap_backward',
+        'fwec': 'endcap_forward',
+        'hr': 'high_resolution',
+        'mantle': 'mantle',
+        'sans': 'sans',
+    }
+    return {
+        InputStreamKey(
+            topic=f'dream_detector_{key}', source_name='dream'
+        ): f'{value}_detector'
+        for key, value in mapping.items()
+    }
+
+
+stream_mapping_dev = make_dev_stream_mapping('dream')
+stream_mapping = StreamMapping(
+    **make_common_stream_mapping_inputs(instrument='dream'),
+    detectors=_make_dream_detectors(),
+)
