+++ src/beamlime/config/instruments/dummy.py
@@ -6,6 +6,10 @@ Detector configuration for a dummy instrument used for development and testing.
 
 import scipp as sc
 
+from beamlime.kafka import InputStreamKey, StreamLUT, StreamMapping
+
+from ._ess import make_common_stream_mapping_inputs, make_dev_stream_mapping
+
 detectors_config = {
     'detectors': {
         'Panel 0': {
@@ -16,3 +20,15 @@ detectors_config = {
         },
     },
 }
+
+
+def _make_dummy_detectors() -> StreamLUT:
+    """Dummy detector mapping."""
+    return {InputStreamKey(topic='dummy_detector', source_name='panel_0'): 'panel_0'}
+
+
+stream_mapping_dev = make_dev_stream_mapping('dummy')
+stream_mapping = StreamMapping(
+    **make_common_stream_mapping_inputs(instrument='dummy'),
+    detectors=_make_dummy_detectors(),
+)
