+++ src/beamlime/config/instruments/nmx.py
@@ -2,6 +2,10 @@
 # Copyright (c) 2025 Scipp contributors (https://github.com/scipp)
 import scipp as sc
 
+from beamlime.kafka import InputStreamKey, StreamLUT, StreamMapping
+
+from ._ess import make_common_stream_mapping_inputs, make_dev_stream_mapping
+
 # TODO Unclear if this is transposed or not. Wait for updated files.
 dim = 'detector_number'
 sizes = {'x': 1280, 'y': 1280}
@@ -27,3 +31,25 @@ detectors_config = {
         },
     },
 }
+
+
+def _make_nmx_detectors() -> StreamLUT:
+    """
+    NMX detector mapping.
+
+    Input keys based on
+    https://confluence.ess.eu/display/ECDC/Kafka+Topics+Overview+for+Instruments
+    """
+    return {
+        InputStreamKey(
+            topic=f'nmx_detector_p{panel}', source_name='nmx'
+        ): 'nmx_detector'
+        for panel in range(3)
+    }
+
+
+stream_mapping_dev = make_dev_stream_mapping('nmx')
+stream_mapping = StreamMapping(
+    **make_common_stream_mapping_inputs(instrument='nmx'),
+    detectors=_make_nmx_detectors(),
+)
